Support interface libraries
CMake complains
"INTERFACE_LIBRARY targets may only have whitelisted properties.  The
property "INCLUDE_DIRECTORIES" is not allowed."
if it is tried to get `INCLUDE_DIRECTORIES` property from an interface
target.

In order to find the include directories of such targets the
`INTERFACE_INCLUDE_DIRECTORIES` property has to be read instead.